게시판만들기(interceptor, 로그인아이디 email삽입, 엑셀파일 다운로드)

+++ sts_workspace/hello-spring/src/main/java/com/hello/forum/bbs/web/BoardControllerr.java
@@ -1,22 +1,36 @@
 package com.hello.forum.bbs.web;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.util.List;
+
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
+import org.apache.poi.ss.usermodel.Workbook;
+import org.apache.poi.xssf.streaming.SXSSFWorkbook;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttribute;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.hello.forum.bbs.service.BoardService;
 import com.hello.forum.bbs.vo.BoardListVO;
 import com.hello.forum.bbs.vo.BoardVO;
 import com.hello.forum.beans.FileHandler;
+import com.hello.forum.member.vo.MemberVO;
 import com.hello.forum.utils.ValidationUtils;
+
+import jakarta.servlet.http.HttpSession;
 
 //import jakarta.validation.Valid;
 
@@ -49,15 +63,23 @@ public class BoardControllerr {
 		return "board/boardlist";
 	}
 	
-	/**
-	 * 게시글 작성 페이지를 보여주는 URL
-	 * @return
-	 */
-	@GetMapping("/board/write") //브라우저에서 링크를 클릭, 브라우저 url을 직접 입력
-	public String viewBoardWritePage(){
+//	/**
+//	 * 게시글 작성 페이지를 보여주는 URL
+//	 * @return
+//	 */
+//	@GetMapping("/board/write") //브라우저에서 링크를 클릭, 브라우저 url을 직접 입력
+//	public String viewBoardWritePage(){
+//		return "board/boardwrite";
+//	}
+	
+	@GetMapping("/board/write")
+	public String viewBoardWritePage(HttpSession session) {
+		MemberVO memberVO = (MemberVO) session.getAttribute("_LOGIN_USER_");
+		if(memberVO == null) {
+			return "redirect:/member/login";
+		}
 		return "board/boardwrite";
 	}
-	
 	
 	//페이지를 보여주는 url :view~
 	// 처리를 하는 url : do~
@@ -98,8 +120,9 @@ public class BoardControllerr {
 //			BindingResult bindingResult, 
 			//Bindingresult : @Valid에 의해 실행된 파라미터 검사(NotEmpty, Email, Size, Min, Max)의 결과
 			@RequestParam MultipartFile file,
-			Model model
-			) {
+			@SessionAttribute("_LOGIN_USER_") MemberVO memberVO, Model model) {
+			
+			{
 		
 		/*
 		* Servlet Like
@@ -129,9 +152,8 @@ public class BoardControllerr {
 		
 		//수동검사 시작 (체크해야하는 파라미터 개수만큼 적어줌)
 		boolean isNotEmptySubject = ValidationUtils.notEmpty(boardVO.getSubject());
-		boolean isNotEmptyEmail = ValidationUtils.notEmpty(boardVO.getEmail());
+		
 		boolean isNotEmptyContent = ValidationUtils.notEmpty(boardVO.getContent());
-		boolean isEmailFormat = ValidationUtils.email(boardVO.getEmail());
 		
 		if(! isNotEmptySubject) {
 			//제목을 입력하지 않았다면
@@ -140,12 +162,6 @@ public class BoardControllerr {
 			return "board/boardwrite";
 		}
 		
-		if(! isNotEmptyEmail) {
-			//이메일을 입력하지 않았다면
-			model.addAttribute("errorMessage", "이메일은 필수입력값 입니다.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardwrite";
-		}
 		
 		if(! isNotEmptyContent) {
 			//내용을 입력하지 않았다면
@@ -153,13 +169,8 @@ public class BoardControllerr {
 			model.addAttribute("boardVO", boardVO);
 			return "board/boardwrite";
 		}
+		boardVO.setEmail(memberVO.getEmail()); //세션에 있는 이메일을 글등록할때 넣어라(굳이 이메일 추가입력 안해도 됨)
 		
-		if(! isEmailFormat) {
-			// 이메일을 이메일 형태로 입력하지 않았다면
-			model.addAttribute("errorMessage", "이메일을 올바른 형태로 작성해주세요.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardwrite";
-		}
 		
 		
 		
@@ -176,7 +187,7 @@ public class BoardControllerr {
 		// 명령을 받은 브라우저는 /board/list로 url을 이동시킨다
 		// 그 이후에 /board/list로 브라우저가 요청을 하게되면
 		// 스프링 컨트롤러에서 /board/list URL에 알맞은 처리를 진행한다
-		return "redirect:/board/list";
+		return "redirect:/board/list";}
 	}
 	
 	// browser에서 url을 http://localhost:8080/board/view?id=1  <-- 나쁘지 않은 방법
@@ -199,11 +210,16 @@ public class BoardControllerr {
 		return "board/boardview";
 	}
 	
+	
 	@GetMapping("/board/modify/{id}") //board/modify/1 <-- id변수의 값은 1
-	public String viewBoardModifyPade( @PathVariable int id, Model model) {
+	public String viewBoardModifyPade( @PathVariable int id, Model model,
+			@SessionAttribute("_LOGIN_USER_") MemberVO memberVO) {
 		//1. 전달받은 id의값으로 게시글을 조회한다
 		BoardVO boardVO = this.boardService.getOneBoard(id, false);
 		
+		if (!memberVO.getEmail().equals(boardVO.getEmail())) {
+			throw new IllegalArgumentException("잘못된 접근입니다.");
+		}
 		//2. 게시글의 정보를 화면에 보내준다
 		model.addAttribute("boardVO", boardVO);
 		
@@ -221,13 +237,17 @@ public class BoardControllerr {
 	public String doBoardModify(@PathVariable int id, 
 								 BoardVO boardVO, 
 								 @RequestParam MultipartFile file,
-								 Model model) {
+								 Model model,
+								 @SessionAttribute("_LOGIN_USER_")MemberVO memberVO) {
+		BoardVO originalBoardVO = this.boardService.getOneBoard(id, false);
+		
+		if (!originalBoardVO.getEmail().equals(memberVO.getEmail())) {
+			throw new IllegalArgumentException("잘못된 접근입니다.");
+		}
 		
 		//수동검사 시작 (체크해야하는 파라미터 개수만큼 적어줌)
 		boolean isNotEmptySubject = ValidationUtils.notEmpty(boardVO.getSubject());
-		boolean isNotEmptyEmail = ValidationUtils.notEmpty(boardVO.getEmail());
 		boolean isNotEmptyContent = ValidationUtils.notEmpty(boardVO.getContent());
-		boolean isEmailFormat = ValidationUtils.email(boardVO.getEmail());
 		
 		if(! isNotEmptySubject) {
 			//제목을 입력하지 않았다면
@@ -236,12 +256,6 @@ public class BoardControllerr {
 			return "board/boardmodify";
 		}
 		
-		if(! isNotEmptyEmail) {
-			//이메일을 입력하지 않았다면
-			model.addAttribute("errorMessage", "이메일은 필수입력값 입니다.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardmodify";
-		}
 		
 		if(! isNotEmptyContent) {
 			//내용을 입력하지 않았다면
@@ -250,18 +264,11 @@ public class BoardControllerr {
 			return "board/boardmodify";
 		}
 		
-		if(! isEmailFormat) {
-			// 이메일을 이메일 형태로 입력하지 않았다면
-			model.addAttribute("errorMessage", "이메일을 올바른 형태로 작성해주세요.");
-			model.addAttribute("boardVO", boardVO);
-			return "board/boardmodify";
-		}
-		
-		
-		
+	
 		// Command Object에는 전달된 ID가 없으므로
 		// PathVariable로 전달된 ID를 셋팅해준다.
 		boardVO.setId(id);
+//		boardVO.setEmail(memberVO.getEmail());
 		
 		boolean isUpdatedSuccess = this.boardService.updateOneBoard(boardVO, file);
 		
@@ -295,7 +302,13 @@ public class BoardControllerr {
 	
 	//삭제 - 이 게시글을 지워라
 	@GetMapping("/board/delete/{id}")
-	public String doDeleteBoard(@PathVariable int id) {
+	public String doDeleteBoard(@PathVariable int id,
+								@SessionAttribute("_LOGIN_USER_") MemberVO memberVO) {
+		
+		BoardVO originalBoardVO = this.boardService.getOneBoard(id, false);
+		if (!originalBoardVO.getEmail().equals(memberVO.getEmail())) {
+			throw new IllegalArgumentException("잘못된 접근입니다.");
+		}
 		
 		//삭제 요청
 		boolean isDeletedSuccess = this.boardService.deleteOneBoard(id);
@@ -329,6 +342,101 @@ public class BoardControllerr {
 		//첨부된파일이 있을경우엔 파일을 사용자에게 보내준다(Download)
 		return this.fileHandler.download(boardVO.getOriginFileName(), boardVO.getFileName());
 	}
+	
+	//excel 다운로드
+	@GetMapping("/board/excel/download")
+	public ResponseEntity<Resource> downloadExcelFile(){
+		
+		//(엑셀 파일을 쓸) 모든게시글 조회
+		BoardListVO boardListVO = boardService.getAllBoard();
+		//XLSX문서 만들기
+		Workbook workbook = new SXSSFWorkbook(-1);
+		//엑셀시트 만들기
+		Sheet sheet = workbook.createSheet("게시글 목록");
+		// 행 만들기
+		Row row = sheet.createRow(0);
+		// 타이틀 만들기
+		Cell cell = row.createCell(0);
+		cell.setCellValue("번호");
+		
+		cell = row.createCell(1);
+		cell.setCellValue("제목");
+		
+		cell = row.createCell(2);
+		cell.setCellValue("첨부파일명");
+		
+		cell = row.createCell(3);
+		cell.setCellValue("작성자이메일");
+		
+		cell = row.createCell(4);
+		cell.setCellValue("조회수");
+		
+		cell = row.createCell(5);
+		cell.setCellValue("등록일");
+		
+		cell = row.createCell(6);
+		cell.setCellValue("수정일");
+		
+		//데이터 행 만들고 쓰기
+		List<BoardVO> boardList = boardListVO.getBoardList();
+		int rowIndex = 1;
+		for(BoardVO boardVO : boardList) {
+			row = sheet.createRow(rowIndex);
+			cell = row.createCell(0);
+			cell.setCellValue(""+boardVO.getId());
+			
+			cell = row.createCell(1);
+			cell.setCellValue(""+boardVO.getSubject());
+			
+			cell = row.createCell(2);
+			cell.setCellValue(""+boardVO.getOriginFileName());
+			
+			cell = row.createCell(3);
+			cell.setCellValue(""+boardVO.getEmail());			
+			
+			cell = row.createCell(4);
+			cell.setCellValue(""+boardVO.getViewCnt());
+			
+			cell = row.createCell(5);
+			cell.setCellValue(""+boardVO.getCrtDt());
+			
+			cell = row.createCell(6);
+			cell.setCellValue(""+boardVO.getMdfyDt());
+			
+			rowIndex += 1;
+		}
+		
+		//엑셀파일 만들기
+		File storedFile = fileHandler.getStoredFile("게시글_목록.xlsx");
+		OutputStream os = null;
+		try {
+			os = new FileOutputStream(storedFile);
+			workbook.write(os);
+		}catch (IOException e) {
+			throw new IllegalArgumentException("엑셀파일을 만들수 없습니다.");
+		} finally {
+			try {
+				workbook.close();
+			} catch (IOException e) {
+				
+			}
+			if (os != null) {
+				try {
+					os.flush();
+				}catch (IOException e) {
+				}
+				try {
+					os.close();
+				}catch (IOException e) {
+				}
+			}
+		}
+		
+		return this.fileHandler.download("게시글_목록.xlsx", "게시글_목록.xlsx");
+//		return this.fileHandler.download("게시글_목록.xlsx", storedFile.getName());
+
+	}
+
 	
 }
